--- .env
@@ -1,8 +1,0 @@
-// DO NOT SHARE THIS FILE WITH ANYONE
-
-CLIENT_ID=""
-TOKEN=""
-GUILD_ID=""
-CAPTAIN_IDS=""
-DATABASE_NAME=""
-DATABASE_URL=""
