+++ src/events/client-ready.js
@@ -4,6 +4,5 @@ module.exports = {
 	execute(client) {
 		const { user } = client
 		console.log(require('chalk').green(`Discord client ready! Logged in as ${user.username} - ID: ${user.id}`))
-		require(`../firestore/init.js`).execute()
 	}
 }
