--- src/firestore/serviceAccountKey.json
@@ -1,14 +1,0 @@
-{
-  "_comment": "DO NOT SHARE THIS FILE WITH ANYONE",
-
-  "type": "",
-  "project_id": "",
-  "private_key_id": "",
-  "private_key": "",
-  "client_email": "",
-  "client_id": "",
-  "auth_uri": "",
-  "token_uri": "",
-  "auth_provider_x509_cert_url": "",
-  "client_x509_cert_url": ""
-}
